Rename isMatchToPncAdjudicationAndDisposals to areResultsMatchingPncAdjudicationAndDisposals
